msm: mdss: refactor pipe list management
A pipe can only belong to either used list or cleanup list, to ensure
this refactor the list management by having only a single list head in
pipe structure. Also add a better name for the lock that should be used
for ensuring atomic access to these lists to prevent incorrect use of
these locks.

This also ensures that a pipe can only be freed on the fb device it was
allocated and prevents unexpected behavior when this is not the case.

Change-Id: Ib1c8ffefe51d837c2b30aa3734d37fee3c54cf1b
Signed-off-by: Adrian Salido-Moreno <adrianm@codeaurora.org>

Conflicts:
	drivers/video/msm/mdss/mdss_mdp_overlay.c