msm: mdss: add check to avoid fudge factor for non real time interfaces
Current driver checks if some video interface is connected
in order to apply a fudge factor. This is expected for real
time interfaces, but non real time interfaces should not
account for this extra bandwidth. This fix adds a check
to make sure that for non real time interfaces the fudge
factor does not get applied.

Change-Id: Ib01d12f4e7749f4e083fa09f6392896313978d72
Signed-off-by: Ingrid Gallardo <ingridg@codeaurora.org>
Signed-off-by: Nirmal Abraham <nabrah@codeaurora.org>