msm: mdss: disable read pointer interrupt during suspend always
During suspend, for command mode panels, we currently disable the
read pointer interrupt and set rdptr_enabled variable to '0' during
timeout cases only if panel is in bad state. We need to perform this
independent of panel state since the DSI clocks needs to be disabled
during suspend always.

Change-Id: I6363fa5a260d26c07b6eff508982c91c51713f29
Signed-off-by: Padmanabhan Komanduru <pkomandu@codeaurora.org>

Conflicts:
	drivers/video/msm/mdss/mdss_mdp_intf_cmd.c